Clean up bot a little
